Revert "add header for cart.jsp"
This reverts commit f990ec5fd238500d374367945bf671480c29da6c.
