--- src/java/Controller/BillServlet.java
@@ -1,128 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
- */
-
-package Controller;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-
-import dal.AccountDAO;
-import dal.OrderDAO;
-import dal.OrderDetailDAO;
-import dal.ProductDAO;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.util.ArrayList;
-import java.util.List;
-import model.OrderDetail;
-import model.Product;
-import model.Bill;
-
-/**
- *
- * @author ASUS PC
- */
-@WebServlet(name = "BillServlet", urlPatterns = { "/invoice" })
-public class BillServlet extends HttpServlet {
-
-    /**
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
-     * methods.
-     * 
-     * @param request  servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException      if an I/O error occurs
-     */
-    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        try (PrintWriter out = response.getWriter()) {
-            /* TODO output your page here. You may use following sample code. */
-            out.println("<!DOCTYPE html>");
-            out.println("<html>");
-            out.println("<head>");
-            out.println("<title>Servlet BillServlet</title>");
-            out.println("</head>");
-            out.println("<body>");
-            out.println("<h1>Servlet BillServlet at " + request.getContextPath() + "</h1>");
-            out.println("</body>");
-            out.println("</html>");
-        }
-    }
-
-    // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the
-    // + sign on the left to edit the code.">
-    /**
-     * Handles the HTTP <code>GET</code> method.
-     * 
-     * @param request  servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException      if an I/O error occurs
-     */
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        Cookie[] cookies = request.getCookies();
-        OrderDAO orderDAO = new OrderDAO();
-        ProductDAO productDAO = new ProductDAO();
-        List<Bill> listbill = new ArrayList<>();
-        int accountID = -1;
-
-        // get id from cookie
-        for (Cookie cookie : cookies) {
-            if (cookie.getName().equals("accountID")) {
-                try {
-                    accountID = Integer.parseInt(cookie.getValue());
-                } catch (Exception e) {
-                    System.out.println(e);
-                }
-            }
-        }
-
-        // get Order from database
-        List<Integer> listOrderID = orderDAO.getAllOrderIDByAccountID(accountID);
-
-        for (int orderID : listOrderID) {
-            // get list order detail from database
-            List<Product> listProduct = productDAO.getListProductByOrderID(orderID);
-            listbill.add(new Bill(listProduct));
-        }
-
-        request.setAttribute("data", listbill);
-        request.getRequestDispatcher("bill.jsp").forward(request, response);
-
-    }
-
-    /**
-     * Handles the HTTP <code>POST</code> method.
-     * 
-     * @param request  servlet request
-     * @param response servlet response
-     * @throws ServletException if a servlet-specific error occurs
-     * @throws IOException      if an I/O error occurs
-     */
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        processRequest(request, response);
-    }
-
-    /**
-     * Returns a short description of the servlet.
-     * 
-     * @return a String containing servlet description
-     */
-    @Override
-    public String getServletInfo() {
-        return "Short description";
-    }// </editor-fold>
-
-}
