+++ web/styles/cart.css
@@ -1,4 +1,5 @@
 body {
+    margin-top: 20px;
     background: #eee;
 }
