+++ web/styles/invoice.css
@@ -14,10 +14,6 @@ body {
 
 .menu {
     background-color: #ffffff;
-    height: 280px;
-    border: 1px;
-    background: white;
-    border-radius: 5px;
 }
 
 .product {
@@ -37,22 +33,14 @@ body {
 }
 
 .second-product {
-    /* display: flex; */
-    /* margin: 0px;
+    display: flex;
+    margin: 0px;
     justify-content: space-between;
-    align-items: center; */
+    align-items: center;
 }
 
 .text {
     text-align: right;
-}
-
-.text .right {
-    margin-left: 10px;
-}
-
-.text p {
-    display: inline;
 }
